Fixed issue with Generating TreeView based on HierarchyID
We encountered that recursion method to find parent record to insert child record was having issues. We just were using Level to find parent but then we had to use HierarchyId.ToString() to verify that we are looking at correct parent.

This works only when we have unique HierarchyID.ToString() in the system for Itemz.